removed junk and renamed stuff
